+++ app/controllers/form_controller.rb
@@ -107,7 +107,7 @@ class FormController < ApplicationController
   end
 
   def check_trade_count_limit
-    if current_user.has_trade_award?
+    if current_account.has_trade_award?
       redirect_to dashboard_url, flash: { alert: 'You can not submit more than one trade form' }
     end
   end
