+++ app/models/account.rb
@@ -6,7 +6,10 @@ class Account < ActiveRecord::Base
   validates :owner, presence: true
 
   def collaborators_without(user)
-    users.excluding(user)
-         .by_email
+    users.excluding(user).by_email
+  end
+
+  def has_trade_award?
+    form_answers.where(award_type: 'trade').any?
   end
 end
