+++ app/views/shared/_application_awards_status.html.slim
@@ -17,7 +17,7 @@
 /* action:    Download PDF of your application
 
 ul.applications-list
-  - unless current_user.has_trade_award?
+  - unless current_account.has_trade_award?
     li
       h3
         ' International Trade Award
